style: rename modules with factory functions
